Deploy flyway migrations from CircleCI
